+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add `is_*` methods for `Value`, `ValueRef` and `NumberRef`.
+
 ## [0.1.2] - 2023-10-30
 
 ### Added